refactor(kitchen pos) : PurgomalumClient 와 구현체 common 패키지로 분리
