+++ src/main/java/kitchenpos/common/external/DefaultPurgomalumClient.java
@@ -1,6 +1,5 @@
-package kitchenpos.menu.infrastructure.external;
+package kitchenpos.common.external;
 
-import kitchenpos.menu.application.external.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
