+++ src/main/java/kitchenpos/common/external/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.application.external;
+package kitchenpos.common.external;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
