+++ src/main/java/kitchenpos/order/common/model/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.order.model;
+package kitchenpos.order.common.model;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
