+++ src/main/java/kitchenpos/order/common/model/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.order.model;
+package kitchenpos.order.common.model;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
