+++ src/main/java/kitchenpos/order/common/ui/OrderRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.order.order.ui;
+package kitchenpos.order.common.ui;
 
-import kitchenpos.order.order.application.OrderService;
-import kitchenpos.order.order.model.Order;
+import kitchenpos.order.common.application.OrderService;
+import kitchenpos.order.common.model.Order;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
