+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -7,7 +7,7 @@ import kitchenpos.menu.domain.model.MenuProduct;
 import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.product.domain.model.Product;
 import kitchenpos.product.domain.repository.ProductRepository;
-import kitchenpos.menu.application.external.PurgomalumClient;
+import kitchenpos.common.external.PurgomalumClient;
 import kitchenpos.menu.infrastructure.persistence.InMemoryMenuGroupRepository;
 import kitchenpos.menu.infrastructure.persistence.InMemoryMenuRepository;
 import kitchenpos.product.infrastructure.InMemoryProductRepository;
