+++ src/test/java/kitchenpos/order/order/application/OrderServiceTest.java
@@ -1,15 +1,16 @@
 package kitchenpos.order.order.application;
 
+import kitchenpos.order.common.application.OrderService;
 import kitchenpos.order.order.infrastructure.external.FakeKitchenridersClient;
 import kitchenpos.menu.infrastructure.persistence.InMemoryMenuRepository;
 import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.order.order.model.Order;
-import kitchenpos.order.order.model.OrderLineItem;
-import kitchenpos.order.order.repository.OrderRepository;
-import kitchenpos.order.order.model.OrderStatus;
+import kitchenpos.order.common.model.Order;
+import kitchenpos.order.common.model.OrderLineItem;
+import kitchenpos.order.common.repository.OrderRepository;
+import kitchenpos.order.common.model.OrderStatus;
 import kitchenpos.order.eatin.model.OrderTable;
-import kitchenpos.order.order.repository.OrderTableRepository;
-import kitchenpos.order.order.model.OrderType;
+import kitchenpos.order.common.repository.OrderTableRepository;
+import kitchenpos.order.common.model.OrderType;
 import kitchenpos.order.order.infrastructure.persistence.InMemoryOrderRepository;
 import kitchenpos.order.order.infrastructure.persistence.InMemoryOrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
